backend/requirements.txt
Django==5.1.2
django-cors-headers==4.6.0
django-extensions==3.2.3
django-vite==3.0.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.3
pytest-django==4.9.0
python-decouple==3.8
sqlparse==0.5.1
vt-py==0.18.4
